+++ example/example.dart
@@ -1,8 +1,10 @@
+// ignore_for_file: avoid_print
+
 import 'package:flutter/material.dart';
 import 'package:vmobject/vmobject.dart';
 
 void main() {
-  runApp(MaterialApp(home: HomeView()));
+  runApp(const MaterialApp(home: HomeView()));
 }
 
 class HomeModel extends ViewModel {
@@ -10,13 +12,15 @@ class HomeModel extends ViewModel {
   final Observable<int> a = Observable(0);
   final Observable<double> b = Observable(0);
   final Observable<double?> c = Observable(null);
-  final Observable<String> d = Observable('');
+  final Observable<String> d = Observable(
+    '',
+    validator: ValidatorLeast([ValidatorRequired(), ValidatorEmail()]),
+  );
   final Observable<DateTime> e = Observable(DateTime(2021, 1, 4));
   final Observable<DateTime?> f = Observable(null);
   final Observable<String> g = Observable('');
   final Observable<DateTime> h = Observable(DateTime.now());
-  final Observable<int?> i = Observable(null)
-    ..isValid.validator = ValidatorRequired();
+  final Observable<int?> i = Observable(null, validator: ValidatorRequired());
 
   @override
   void activate() {
@@ -46,60 +50,59 @@ class HomeView extends ViewWidget<HomeModel> {
   Widget builder(BuildContext context, HomeModel model) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('My App'),
+        title: const Text('My App'),
       ),
       body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          mainAxisSize: MainAxisSize.min,
+        child: ListView(
           children: <Widget>[
-            Text(
+            const Text(
               'You have pushed the button this many times:',
             ),
             ObserverWidget(
-                observable: model.count,
-                builder: (context, int value) => Text(
-                      '$value',
-                      style: Theme.of(context).textTheme.headline4,
-                    )),
+              observable: model.count,
+              builder: (context, int value) => Text(
+                '$value',
+                style: Theme.of(context).textTheme.headline4,
+              ),
+            ),
             ObserverFormField(
               observable: model.a,
-              decoration: InputDecoration(labelText: 'Int'),
+              decoration: const InputDecoration(labelText: 'Int'),
             ),
             ObserverFormField(
               observable: model.b,
-              decoration: InputDecoration(labelText: 'Double'),
+              decoration: const InputDecoration(labelText: 'Double'),
             ),
             ObserverFormField(
               observable: model.c,
-              decoration: InputDecoration(labelText: 'Double?'),
+              decoration: const InputDecoration(labelText: 'Double?'),
             ),
             ObserverFormField(
               observable: model.d,
-              decoration: InputDecoration(labelText: 'String'),
+              decoration: const InputDecoration(labelText: 'Email'),
             ),
             ObserverFormField(
               observable: model.e,
-              decoration: InputDecoration(labelText: 'DateTime'),
+              decoration: const InputDecoration(labelText: 'DateTime'),
             ),
             ObserverFormField(
               observable: model.f,
-              decoration: InputDecoration(labelText: 'DateTime?'),
+              decoration: const InputDecoration(labelText: 'DateTime?'),
             ),
             ObserverFormField(
               observable: model.g,
-              decoration: InputDecoration(labelText: 'Mask'),
+              decoration: const InputDecoration(labelText: 'Mask'),
               formatter: MaskTextFormatter(mask: '0000-####-####'),
             ),
             ObserverFormField(
               observable: model.h,
-              decoration: InputDecoration(labelText: 'Full DateTime'),
+              decoration: const InputDecoration(labelText: 'Full DateTime'),
               formatter: DateTimeFormatter<DateTime>(
                   type: DateTimeFormatterType.dateFullTime),
             ),
             ObserverFormField(
               observable: model.i,
-              decoration: InputDecoration(labelText: 'Required'),
+              decoration: const InputDecoration(labelText: 'Required number'),
             ),
           ],
         ),
@@ -109,7 +112,7 @@ class HomeView extends ViewWidget<HomeModel> {
           model.add();
         },
         tooltip: 'Increment',
-        child: Icon(Icons.add),
+        child: const Icon(Icons.add),
       ),
     );
   }